Fix documentation (and comments) for colour usage: client, session
Colour usage is documented at various places in the repository,
including the README.md file, the main powerline.kak script, and the
colour themes themselves. Looking at those various files, it appears
that the description for a few colours is incorrect: the foreground and
background colours are swapped for the client module in several cases,
and for the session module in all cases.

We could address either by updating the documentation, or by changing
the colours in use to reflect the current documentation. The latter
option would likely introduce colour changes for the users of the
existing themes, so instead we simply update the documentation and do
not change how the colours are used.

Fixes: bb4c9bf0f04c ("reorganize colors")